BUG: adcl returns int64

+++ spectre/factors/feature.py
@@ -50,8 +50,8 @@ class AdvanceDeclineRatio(CrossSectionFactor):
     win = 1
 
     def compute(self, returns):
-        advancing = nansum(returns > 0, dim=1)
-        declining = nansum(returns < 0, dim=1)
+        advancing = nansum(returns > 0, dim=1).float()
+        declining = nansum(returns < 0, dim=1).float()
         ratio = (advancing / declining).unsqueeze(-1)
         return ratio.expand(ratio.shape[0], returns.shape[1])
